Configure MongoDatabaseSettings
To be able to use the Options pattern in the NoSql repositories we
should configure the options itself, in this case MongoDatabaseSettings
that has the connection string, database and collections names.
To able to use the Options pattern in .net we should configure the
optiosn itself in this case is the MongoDatabaseSettings.