Refactor Player LeaveTheTeam
The fact that a playe can exit without a Team makes their relationship
optional, the problem is that when a player leaves the team the TeamId
is initialized to an empty GUID but the database doesn't know how to
deal with this and tries to relate the player to a Team that has an
empty GUID as id and of course this throws an exception.

So we changed that and now when a player leaves the team we initialize
the TeamId to null.