Refactor entities state
The states of the entities were very similar, for example, the methods
of setting the current entity or loading the entities were repeated.

We abstract all that logic into a base class.

For this we had to identify something in common in the entities and that
was their id, so we made the API responses inherit IEntity.

With this we can work on the base class and then create concrete states
to add unique functionality to each entity.